+++ klaw-gui/src/widgets/array_editor.rs
@@ -103,5 +103,4 @@ mod tests {
 
         assert_eq!(vec[0], "value");
     }
-
 }